master branch test01

+++ src/components/Modal.js
@@ -5,19 +5,24 @@ import { useRouteMatch } from 'react-router-dom';
 const Modal = ({ rweetObj, setRweetObj, setOnModal, isOwner, setUpdated }) => {
 	const [newRweet, setNewRweet] = useState(rweetObj.text);
 	const [editing, setEditing] = useState(false);
+	const [doubleSubmit, setDoubleSubmit] = useState(false);
 	let isProfile = useRouteMatch('/profile');
 
 	const onDeleteClick = async () => {
-		const ok = window.confirm('Are you sure you want to delete this rweet?');
-		if (ok) {
-			await dbService.doc(`rweets/${rweetObj.id}`).delete();
-			if (rweetObj.attachmentUrl !== '') {
-				await storageService.refFromURL(rweetObj.attachmentUrl).delete();
+		if (!doubleSubmit) {
+			setDoubleSubmit(true);
+			const ok = window.confirm('Are you sure you want to delete this rweet?');
+			if (ok) {
+				goOut_modal();
+				await dbService.doc(`rweets/${rweetObj.id}`).delete();
+				if (rweetObj.attachmentUrl !== '') {
+					await storageService.refFromURL(rweetObj.attachmentUrl).delete();
+				}
+				if (isProfile) {
+					setUpdated(Date.now());
+				}
 			}
-			goOut_modal();
-			if (isProfile) {
-				setUpdated(Date.now());
-			}
+			setDoubleSubmit(false);
 		}
 	};
